+++ src/test/kotlin/me/bossm0n5t3r/leetcode/searchInRotatedSortedArray/SearchInRotatedSortedArrayTest.kt
@@ -1,10 +1,10 @@
-package searchInRotatedSortedArray
+package me.bossm0n5t3r.leetcode.searchInRotatedSortedArray
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class SearchInRotatedSortedArrayTest {
-    private val sut = SearchInRotatedSortedArray()
+    private val sut = SearchInRotatedSortedArray.Solution()
 
     private data class Example(
         val nums: IntArray,
